+++ CHANGELOG.md
@@ -18,5 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add `dependabot` version updates
 - Optimize `test` workflow
 - Move prod URL validation and request sending to `__init__.py`
+- Implement tasks API
+- Add tests for tasks API
 
 [Unreleased]: https://github.com/IceBotYT/pynoonlight/compare/v0.1.0..HEAD